backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.0.0
opencv-python>=4.9.0
ultralytics>=8.2.0
numpy>=1.26.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
